requirements.txt
openai
streamlit
python-dotenv
PyPDF2
python-docx
tiktoken
